feat:edit employer routes add industry tags in the profile component
