+++ src/components/company/jobdetails.js
